refactor: 멤버 활동 정보 섹션 컴포넌트 분리
