+++ src/api/endpoint_LEGACY/members/type.ts
@@ -55,17 +55,7 @@ export type ProfileDetail = {
   };
   idealType: string;
   selfIntroduction: string;
-  soptActivities: {
-    generation: number;
-    part: string;
-    team: string | null;
-    projects: {
-      id: number;
-      generation: number;
-      name: string;
-      category: ProjectCategory;
-    }[];
-  }[];
+  soptActivities: SoptActivity[];
   links: MemberLink[];
   projects: MemberProject[];
   careers: Career[];
@@ -73,12 +63,16 @@ export type ProfileDetail = {
   isMine: boolean;
 };
 
-export type Activity = {
-  id: number;
+export type SoptActivity = {
   generation: number;
-  team: string;
   part: string;
-  isProject: boolean;
+  team: string | null;
+  projects: {
+    id: number;
+    generation: number;
+    name: string;
+    category: ProjectCategory;
+  }[];
 };
 
 export type MemberLink = {
